Move this mixin into the admin file instead
